requirements.txt
django
boto3
djangorestframework
groq
python-dotenv